+++ assets/src/edit-story/app/story/actions/useSelectElementById.js
@@ -7,6 +7,7 @@ function useSelectElementById( {
 	setSelectedElementIds,
 } ) {
 	const selectElementById = useCallback( ( id ) => {
+		console.log( 'in-selected ids', id );
 		setSelectedElementIds( [ id ] );
 	}, [ setSelectedElementIds ] );
 	return selectElementById;
